Detached FinePrint config from FinePrint itself.
FinePrint's config file is not accessible to ModuleManager, meaning that
mods which wish to modify it need to overwrite the existing file. While
we anticipate this behaviour will change with KSP 0.90 (when FinePrint
enters core), in the meantime we still need a way to handle it.

This change has FinePrint require a config file, but allows mods to
supply their own. Currently only *recent* releases of RP-0 do this,
and I've updated the NetKAN file to correctly support those.

Happily tested with my own setup, and you can test this right now on
the super secret PaulKAN:

    ckan.exe update --repo=https://github.com/pjf/CKAN-meta/archive/paulkan.zip