FirespitterCore now includes Firespitter.cfg
- Relieves #226
- Brings joy to the world

The only gotcha is if someone already has FSCore installed using the old
metadata, and then tries to install Firespitter, they could end up
missing out on `Firespitter.cfg`. The solution to this would simply be
to reinstalll FirespitterCore (`ckan upgrade FirespitterCore` will do
the right thing, too).

Alternatively, I could bump the version on FSCore just a smidgen, but
then folks who already have the old FS metadata will get a conflict if
they try to install the "update". The solution here is to reinstall
Firespitter (`ckan upgrade Firespitter` works, too).

The *best* way of doing this would be if the CKAN actually respected
version numbers in dependencies, but it doesn't do that yet.

CC'ing in @snjo and @Ippo343, as this has potential to cause mild grief.