Updated EngineIgnitor repack
The repack works just fine in 0.90 as well as 0.25, and remains
identical to the official release, so I've made a new release with
updated AVC info. This PR updates the metadata to use the `find` keyword
for locating the files, removing a rather ugly kludge with a hard-coded
version number.