Fix much metadata
I've no idea why travis is being so grumpy, but running `prove` in the
NetKAN repo revealed a number of metadata errors. This commit doesn't
fix all of them, but it fixes the easy-to-fix ones. In particular:

* A number of documents have been given a spec `v1.2` or `v1.4` header, as they use newer features.
* A number of files have been renamed to match their identifier.

This doesn't fix hard errors, like missing licenses (which might be okay
if the remote data is fine), or identifiers which don't match the CKAN
spec (which maybe we should just update to allow).