build.sh: Evaluate only changes on branch, not absolute diffs
This change instructs git/build.sh to only examine files which have changed on the current branch from its most-recent-ancestor of `origin/master`. This is better than the absolute differences, as it means `origin/master` can change as much as it likes and we don't care.

Tested locally. I'm sure there's an issue number for this somewhere, so feel free to edit this PR to include that. :)