Update SETI-CommunityTechTree.netkan
Hello,
since the SETI mod complex will undergo a lot of changes from 1.0.4 to 1.0.5 and then again to 1.1, with many changes in between and a move to github in general, I m trying to prepare for it.

To be able to respond faster if (when) issues occur (not only with my mod, but with dependencies/recommendations), I try to externalize the metadata from the netkan respository. I took RealismOverhaul as an example for this. Hopefully I did not miss something.

As this is my first externalization attempt, any feedback on the 2 files (the one inside the netkan respository and the linked file in the SETI-CommunityTechTree repository) is very much appreciated.

Best regards,
Yemo