Create SETI-RemoteTech.netkan
So, the next experiment in terms of Github.

Oh, and thank you Olympic1 for the comma notice in SETI-BalanceMod, the config has gotten quite large and this should be the last change before I update the new SETI-BalanceMod. Will look through the netkan files again before doing that.

Thank you very much!