Add glue to index RoverDude's mods from his own metadata
The `netkan` indexer can use `.netkan` files from other sites, but it
needs glue files to know where to fetch them from. This hange adds that
glue (and is a pretty good reference if we want to do this for other
folks, too).

Tested locally with AlcubierreStandalone and USI-SRV, but I don't have
any reason to expect any of the glue here is incorrect.

Related: #2996, #3009.