Renaming for spacedock.info
SETI-BalanceMod from kerbalstuff renamed to SETI-Rebalance for spacedock.info, I ll adjust my metadata as soon as I have the new spacedock id. 
I can not find any dependencies or references within netkan on the former "SETI-BalanceMod" identifier. 
And with kerbalstuff gone, this is the ideal point in time for a name change.