Fixup RSS / KSP version oddity.
The KSP-AVC spec isn't really clear on whether KSP versions should
always have three parts. Our code assumes that it does. The RSS .version
version file doesn't. This results in a file being emitted targetting
KSP "1.1." (with a trailing dot) which causes validation to fail.

This override forces the version to "1.1", which will work with both
1.1.0 and 1.1.1 (which is arriving soon).

For @NathanKell.