create netkan file for davon supply mod on spacedock
Hello I have created this pull request to add this file to NetKAN. A similar file "DavonSupplyMod.kerbalstuff" already exists which this file should replace. The .netkan file itself is tested with the ckan.exe and netkan.exe utilities. I am the owner of the mod this file refers to.

Yesterday a similar pull request was made by me but it did not show up in the history, so I suspect I did something wrong, hence I'm trying once more.