Add PSA-Greenhouse (#4258)
* Add PSA-Greenhouse

See [forum
thread](http://forum.kerbalspaceprogram.com/index.php?/topic/141704-112-psa-greenhouse-for-usi-life-support-v-1123-2016jun13/).

* Fix spec version

This is what I get for doing this on OS X. >.<

* Fix spec version; reshuffle dependencies

USI-LS technically not required but there's little reason to have the
part without it, so it's in recommends. MKS supported but not required.
No benefit/interaction with CLS really so not included.