update in response to reqeusts on PR.
I really don't think of the forum post as the mod's "home page" at all.  The home page is where it's hosted on GitHub, where there's a lot richer set of tools for documenting it than in a forum post.  But, if that's what it takes to get it into CKAN, then okay I'll make the change.