New entry for LaserDist mod I just released (#4249)
* New entry for LaserDist mod I just released

I just released LaserDist on spacedock and github (soon to be Curse also once I finish submitting this PR for NETKAN).  This is the new netkan entry for it.

* update in response to reqeusts on PR.

I really don't think of the forum post as the mod's "home page" at all.  The home page is where it's hosted on GitHub, where there's a lot richer set of tools for documenting it than in a forum post.  But, if that's what it takes to get it into CKAN, then okay I'll make the change.