Update and rename CoatlAerospace.netkan to ProbesPlus.netkan
I looked at the guide and I hope I have the right changes to reflect the change of my mod to the new released version 0.14 . I'd appreciate if someone can help me get my mod updated, or just proofread my code chage. I suggest the name change to "ProbesPlus" because "Coatl Aerospace" is the name of the in-game company for my mods, whereas ProbesPlus is the mod name.