Add PartOverhauls.netkan (#4645)
* Add PartOverhauls.netkan

I am hosting the .zip as the original .zip was 15 times larger because it contained the source. It also seemed to be throttled to 72KB/s, so it took way too long to download. See link for original source:

http://forum.kerbalspaceprogram.com/index.php?/topic/147582-update-12-pre-release-is-here

* Changed ksp_version_min to ksp_version

Since these configs use modules new to 1.2.0 which is still only in prerelease, I'm not going to give it a ksp_version_max of 1.2.99, just in case something changes before release.