ShipEffects update 1 of 2
This is a NetKAN file for a KSP-1.1.3 compatible ShipEffects that is newer than the one presently in the repository.

I am hoping this pull request generates CKAN-meta/ShipEffects/ShipEffects-1.0.5.ckan

After it has been picked up I will do another PR with a new version of ShipEffects.netkan to generate CKAN-meta/ShipEffects/ShipEffects-1.0.7.ckan which is for KSP 1.2.

Confused yet?  I am.