Reduce ProbesPlus dependencies
The upstream zip includes the FireSpitter plugin and resources,
and the DMagic animation plugin, but not their respective parts;
these are apparently the hard dependencies

DMagicOrbitalScience is needed "for some science experiments to work",
but much of the mod functions well without it; demote to recommends.

The firespitter parts do not seem relevant.