Update BlueHawkIndustries.netkan (#4993)
* Update BlueHawkIndustries.netkan

Closes #4977

Very unsure on the version string. I think this will cound as a new epoch? The mod used to follow a 2.1-2.2-...-7.1 etc version scheme but now it's based on version 1.0-1.1-...-1.5 so a new epoch is needed.

This PR does not fix the borked 1.2 version we have indexed in CKAN-meta nor does it fix any other broken versions which might exist in there.

* Jenkins is a cranky old jsonlint