Update InlineBallutes version
Hello, this is my first Pull Request on Github so please be forgiving. I already tried to make a pull request but this is not showing nor in the NetKAN repository nor in my fork so I am doing this new one (the last one I promise) just in case.

I am not the mod author and the mod is already in CKAN but is listed under a very old version 1.2.3 for KSP 1.0.5 opposed to the new one 1.2.7 for KSP 1.2.0. So I decided to write this netkan file and it is installing correctly in my CKAN. I already informed the mod author by writing a post on the release thread on Kerbal Forums. 

Thanks for the good work with CKAN!
Giorgio from Italy