Create PoodsOPMVO.netkan (#5354)
* Create PoodsOPMVO.netkan

.netkan file for Pood's Outer Planets Mod - Visual Overhaul.

I did notice that the Outer Planets Mod is currently listed as KSP Max Version 1.2.1 and my mod states a requirement of 1.2.2. Outer Planets Mod works fine and is compatible with KSP 1.2.2 and should be updated to reflect this. I will raise a separate issue.

* Update PoodsOPMVO.netkan

Install PoodsOPMVO dir to GameData

* Update PoodsOPMVO.netkan

* Update PoodsOPMVO.netkan

* Update PoodsOPMVO.netkan

Case matters.