Make OPM recommend, not require, CBK. (#5536)
Custom Barn Kit is needed only to increase the communication range of the
tracking station in order to reach the outer planets. The same goal can be
accomplished by using mods like Extended Antenna Progression (or just
building a really overpowered relay satellite). Since CBK is not
compatible with some other mods and is not strictly required for gameplay,
it should be recommended, not required.