Create RealAntennas.netkan
First time adding a mod here.  I believe that the format is correct, and the output from running netkan against this looks reasonable to me.  I just published the official release on my repo.  Please let me know if anything needs to be changed or added!

Looking forward, when I want to release a version of the mod specifically targeting KSP 1.8, what is the right process?  In particular, after publishing a new release on my github repo, should I then go through this process to change the netkan version info?