Update ROEngines.netkan
Move .netkan information into here, instead of maintaining in external ROMods repos.

Be more consistent in where ROMods (ROCapsules, etc) stores the netkan info.
Attempt to be more nuanced in depends/recommends/suggests choices.  The hard dependency on RealismOverhaul is very blunt, but many patches here depend on RO's own patching sequence and it's not worth the effort to get away from that.