Sort and fix Spectra NetKAN Provides/conflicts/depends/suggests/recommends
I updated Spectra used this time to fix its NetKAN. I moved minor mods from Recommends to Suggests (added suggests), removed PlanetShine-config from Provides (No longer providing), removed GreenSkullRevived and PlanetShine-config from Conflicts (no longer conflicting), and removed Scatterer-sunflare-default from Depends (no longer depends).

I hope this looks in order, the file did not have suggests before this change, I added it from scratch based on the NetKAN of other mods. The goal is for all the mods in "recommends" to be auto-ticked when you install, and the mods in "suggests" to not.

I also reworked x_netkan_override. This might be set up incorrectly, though it doesn't matter as much as the rest. The current release of Spectra (1.3) is backwards compatible all the way to KSP version 1.3. Is this set up correctly?