Add TransferWindowPlanner - Fork
I forked TWP to make it useful for transfers between planets that have non-zero inclination (by specifying which inclined parking orbit should be used). This is especially important in RSS, but also in stock it can reduce Δv requirements by several hundred m/s for some transfers. 
In the process, I updated it for 1.12.

It would be great to have on CKAN for easier installation, but I want it clear to users that it is a fork and not the original mod by TriggerAu.