Add Infinite Discoveries Effects from SpaceDock (#10090)
* Add Infinite Discoveries Effects from SpaceDock

This is an automated commit on behalf of Sushut

* add dependency on scatterer-sunflare-default

This mod *patches* the sunflare rather than providing one.  It should work with any other sunflare mod (e.g. Dan's) but it will just overwrite and that will be confusing.  If we require the default one things should be more predictable.

---------

Co-authored-by: Sushut <andyduha1981@gmail.com>
Co-authored-by: JonnyOThan <jonnyothan@gmail.com>